Remove warning.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/cataclysm@5953 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_stats.cpp
@@ -190,7 +190,7 @@ void stats_t::analyze()
 
   for ( int i=0; i < num_buckets; i++ )
   {
-    if( i == sim -> divisor_timeline.size() ) break;
+    if( i == (int) sim -> divisor_timeline.size() ) break;
     timeline_dmg[ i ] /= sim -> divisor_timeline[ i ];
   }
